Converted test to jest
Everything is working except for termsStats which is confusing.
