+++ integration-test/example-types/textAndMongoId.js
@@ -1,4 +1,3 @@
-let { expect } = require('chai')
 let types = require('../../types')()
 let Contexture = require('contexture')
 let provider = require('../../src')
@@ -67,8 +66,8 @@ describe('Grouping text and mongoId', () => {
     }
     let result = await process(dsl, { debug: true })
     let response = _.last(result.items).context.response
-    expect(response.totalRecords).to.equal(3)
-    expect(response.results[0]._id.toString()).to.equal(id.toString())
+    expect(response.totalRecords).toBe(3)
+    expect(response.results[0]._id.toString()).toBe(id.toString())
   })
 
   it('should work with populate', async () => {
@@ -115,8 +114,8 @@ describe('Grouping text and mongoId', () => {
     }
     let result = await process(dsl, { debug: true })
     let response = _.last(result.items).context.response
-    expect(response.totalRecords).to.equal(3)
-    expect(response.results[0]._id.toString()).to.equal(id.toString())
-    expect(response.results[0].nextCode.toString()).to.equal(id2.toString())
+    expect(response.totalRecords).toBe(3)
+    expect(response.results[0]._id.toString()).toBe(id.toString())
+    expect(response.results[0].nextCode.toString()).toBe(id2.toString())
   })
 })
