+++ test/example-types/dateHistogram.test.js
@@ -1,12 +1,11 @@
 let _ = require('lodash/fp')
-let { expect } = require('chai')
 let { MongoClient } = require('mongodb')
 let { MongoMemoryServer } = require('mongodb-memory-server')
 let dateHistogram = require('../../src/example-types/dateHistogram')
 
 let aggregate
 
-before(async () => {
+beforeAll(async () => {
   let mongoServer = new MongoMemoryServer()
   let mongoUri = await mongoServer.getConnectionString()
   let conn = await MongoClient.connect(mongoUri, {
@@ -49,7 +48,7 @@ describe('dateHistogram', () => {
         search
       )
 
-      expect(query).eql([
+      expect(query).toEqual([
         {
           $group: {
             _id: {
@@ -81,7 +80,7 @@ describe('dateHistogram', () => {
         },
         { $sort: { year: 1, month: 1, day: 1 } },
       ])
-      expect(result).eql({
+      expect(result).toEqual({
         entries: [
           {
             key: 1580515200000,
